pyproject.toml
[project]
name = "blog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
  "boto3>=1.40.71",
  "celery>=5.4.0",
  "django>=5.2.7",
  "django-environ>=0.12.0",
  "django-solo>=2.4.0",
  "django-storages[s3]>=1.14.6",
  "django-taggit>=6.1.0",
  "markdown>=3.9",
  "markdownify>=0.12.0",
  "pillow>=10.4.0",
  "psycopg>=3.2.12",
  "psycopg2-binary>=2.9.11",
  "python-frontmatter>=1.1.0",
  "redis>=5.0.8",
  "requests>=2.32.5",
]

[dependency-groups]
dev = ["django-debug-toolbar>=6.0.0"]
prod = ["gunicorn>=23.0.0"]
